+++ email/src/account/sync/config.rs
@@ -1,6 +1,8 @@
 use serde::{Deserialize, Serialize};
 use std::path::PathBuf;
 
+use crate::folder::sync::config::FolderSyncStrategy;
+
 #[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct SyncConfig {
@@ -11,4 +13,7 @@ pub struct SyncConfig {
     /// Customize the root directory where the Maildir cache is
     /// saved. Defaults to `$XDG_DATA_HOME/himalaya/<account-name>`.
     pub dir: Option<PathBuf>,
+
+    #[deprecated(since = "0.22.0", note = "moved to FolderConfig::sync::filter")]
+    pub strategy: Option<FolderSyncStrategy>,
 }